+++ cai_causal_graph/__init__.py
@@ -30,7 +30,7 @@ __all__ = [
     'VARIABLE_NAME',
 ]
 
-__version__ = '0.2.9'
+__version__ = '0.2.10'
 
 from cai_causal_graph.causal_graph import CausalGraph, Skeleton
 from cai_causal_graph.time_series_causal_graph import TimeSeriesCausalGraph