+++ changelog.md
@@ -8,6 +8,7 @@
   `cai_causal_graph.time_series_graph.TimeSeriesGraph` class to get the nodes at a given lag, the contemporaneous nodes
   of a node, and the potential contemporaneous parents of a node (inbound directed or unidirected and outbound
   undirected edges) , respectively.
+- General improvements to several `from_*` methods in the `cai_causal_graph.time_series_graph.TimeSeriesGraph` class.
 
 ## 0.2.9